Migration: Change frequency to double in samples table

+++ database/migrations/2016_06_20_015240_change_frequency_to_double_in_samples.php
@@ -13,7 +13,7 @@ class ChangeFrequencyToDoubleInSamples extends Migration
     public function up()
     {
         Schema::table('samples', function (Blueprint $table) {
-            $table->double('frequency')->change();
+            $table->float('frequency')->change();
         });
     }
